+++ database/migrations/2024_06_03_105038_create_ticket_prices_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('ticket_prices', function (Blueprint $table) {
             $table->id();
             $table->string('libelle');
-            $table->double('prix');
+            $table->float('prix');
             $table->string('devise');
             $table->boolean('is_promotion');
             $table->timestamps();
